# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.CreateUserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  dto.LogoutRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.LogoutResponse:
    properties:
      isSuccess:
        type: boolean
    type: object
  dto.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RefreshResponse:
    properties:
      refreshToken:
        type: string
    type: object
  dto.SetClaimsRequest:
    properties:
      claim_ids:
        items:
          type: integer
        type: array
      user_id:
        type: integer
    required:
    - claim_ids
    - user_id
    type: object
  dto.SetClaimsResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for the Knowstack project
  title: Knowstack API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: Checks if the service is alive
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check the liveness of the service
      tags:
      - API Health
  /users/claims:
    post:
      consumes:
      - application/json
      description: Sets claims for a user
      parameters:
      - description: User to set claims for
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SetClaimsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SetClaimsResponse'
      security:
      - BearerAuth: []
      summary: Set claims for a user
      tags:
      - API User
  /users/login:
    post:
      consumes:
      - application/json
      description: Logs in a user
      parameters:
      - description: User to login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
      summary: Login a user
      tags:
      - API User
  /users/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user
      parameters:
      - description: User to logout
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LogoutResponse'
      summary: Logout a user
      tags:
      - API User
  /users/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes a token
      parameters:
      - description: User to refresh
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
      summary: Refresh a token
      tags:
      - API User
  /users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateUserResponse'
      summary: Create a new user
      tags:
      - API User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"